requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
requests
python-dotenv
